Fixed test run status. Renamed components.
