+++ odd-platform-ui/src/components/shared/AlertStatusItem/AlertStatusItemStyles.ts
